small name refactors
